D_all/requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-allauth
django-filter
idna
oauthlib
pycparser
PyJWT
python-dotenv
python3-openid
requests
requests-oauthlib
sqlparse
urllib3
celery
django-apscheduler
redis
